restructure internal methods into classes
