+++ tests/test_api_response.py
@@ -1,7 +1,6 @@
-from bs4 import BeautifulSoup
-from ncsu_courses.util import _get_course_html
 from ncsu_courses.term import Term, Session
-from ncsu_courses.subjects import get_all_subjects
+from ncsu_courses._api import API
+from bs4 import BeautifulSoup
 
 
 def test_api_course_equals_expected(expected_f22_csc_course_html):
@@ -9,14 +8,17 @@ def test_api_course_equals_expected(expected_f22_csc_course_html):
     Assert that the API's returned course HTML equals the expected
     course HTML.
     '''
+    subject = "CSC - Computer Science"
+    term = Term(22, Session.Fall)
+    course_html = API.get_course_html(subject, term)
+
+    course_soup = BeautifulSoup(course_html, 'html.parser')
+
     expected_course_soup = BeautifulSoup(
         expected_f22_csc_course_html, 'html.parser')
 
-    subject = "CSC - Computer Science"
-    term = Term(22, Session.Fall)
-    course_html = _get_course_html(subject, term)
-    course_soup = BeautifulSoup(course_html, 'html.parser')
-
+    # asserting equality is quicker between BeautifulSoup encoded objects than
+    # the raw html strings
     assert expected_course_soup.encode_contents() == course_soup.encode_contents()
 
 
@@ -26,6 +28,6 @@ def test_api_subjects_equals_expected(expected_f22_subjects):
     array of subjects.
     '''
     term = Term(22, Session.Fall)
-    f22_subjects = str(get_all_subjects(term))
+    f22_subjects = str(API.get_subjects_list(term))
 
     assert expected_f22_subjects == f22_subjects
